Refactor database.py to update original_publication_year data type

+++ services/REST/app/data/database.py
@@ -15,7 +15,7 @@ def init_db():
             author VARCHAR(255),
             average_rating FLOAT,
             isbn VARCHAR(13),
-            original_publication_year INT
+            original_publication_year FLOAT
         );
     """)
     conn.commit()
